requirements.txt
alembic==1.17.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
Authlib==1.6.6
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ordered-set==4.1.0
p==1.5.0
packaging==25.0
passlib==1.7.4
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
qrcode==8.2
redis==7.1.0
requests==2.32.5
rich==14.2.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.3
wrapt==2.0.1
